Rename segmentation survey to assessment
